refactor(auth): collapse both direct seams into shadowAuth
