+++ web/src/pages/api/public/integrations/blob-storage/index.ts
@@ -18,7 +18,7 @@ import {
 import { upsertBlobStorageIntegration } from "@/src/features/blobstorage-integration/service";
 import { resolveExportSource } from "@/src/features/analytics-integrations/server/exportSource";
 import { auditLog } from "@/src/features/audit-logs/auditLog";
-import { verifyOrgAuth } from "@/src/features/auth/policy/verifyOrgAuth";
+import { shadowAuth } from "@/src/features/public-api/server/shadowAuth";
 
 /** orgKeyRequired is the 403 body when a non-organization key hits a blob-storage endpoint. */
 const orgKeyRequired =
@@ -196,10 +196,14 @@ async function handleUpsertBlobStorageIntegration(
 async function authorizeBlobStorageRequest(
   req: NextApiRequest,
 ): Promise<ApiAccessScope> {
-  const authCheck = await verifyOrgAuth({ req, action: "projects:read" });
-  if (!authCheck.validKey) {
-    if (authCheck.status === 401) {
-      throw new UnauthorizedError(authCheck.error);
+  const authCheck = await shadowAuth({
+    req,
+    action: "projects:read",
+    allowedAccessLevels: ["organization"],
+  });
+  if (!authCheck.success) {
+    if (authCheck.error.httpCode === 401) {
+      throw new UnauthorizedError(authCheck.error.message);
     }
     throw new ForbiddenError(orgKeyRequired);
   }
