+++ web/src/pages/api/public/organizations/apiKeys/index.ts
@@ -2,7 +2,7 @@ import { cors, runMiddleware } from "@/src/features/public-api/server/cors";
 import { logger } from "@langfuse/shared/src/server";
 import { RateLimitService } from "@/src/features/public-api/server/RateLimitService";
 import { handleGetApiKeys } from "@/src/ee/features/admin-api/server/organizations/apiKeys";
-import { verifyOrgAuth } from "@/src/features/auth/policy/verifyOrgAuth";
+import { shadowAuth } from "@/src/features/public-api/server/shadowAuth";
 
 import { type NextApiRequest, type NextApiResponse } from "next";
 import { hasEntitlementBasedOnPlan } from "@/src/features/entitlements/server/hasEntitlement";
@@ -27,13 +27,15 @@ export default async function handler(
   }
 
   // CHECK AUTH
-  const authCheck = await verifyOrgAuth({
+  const authCheck = await shadowAuth({
     req,
     action: "organization:CRUD_apiKeys",
+    allowedAccessLevels: ["organization"],
   });
-  if (!authCheck.validKey) {
-    return res.status(authCheck.status).json({
-      error: authCheck.status === 403 ? orgKeyRequired : authCheck.error,
+  if (!authCheck.success) {
+    const status = authCheck.error.httpCode;
+    return res.status(status).json({
+      error: status === 403 ? orgKeyRequired : authCheck.error.message,
     });
   }
   // END CHECK AUTH
