+++ web/src/pages/api/public/projects/[projectId]/apiKeys/[apiKeyId].ts
@@ -8,7 +8,7 @@ import {
   handleDeleteApiKey,
 } from "@/src/ee/features/admin-api/server/projects/projectById/apiKeys/apiKeyById";
 import { hasEntitlementBasedOnPlan } from "@/src/features/entitlements/server/hasEntitlement";
-import { verifyOrgAuth } from "@/src/features/auth/policy/verifyOrgAuth";
+import { shadowAuth } from "@/src/features/public-api/server/shadowAuth";
 
 /** orgKeyRequired is the 403 body when a non-organization key hits an organization endpoint. */
 const orgKeyRequired =
@@ -34,13 +34,15 @@ export default async function handler(
     const { projectId, apiKeyId } = params;
 
     // CHECK AUTH
-    const authCheck = await verifyOrgAuth({
+    const authCheck = await shadowAuth({
       req,
       action: "apiKeys:CUD",
+      allowedAccessLevels: ["organization"],
     });
-    if (!authCheck.validKey) {
-      return res.status(authCheck.status).json({
-        message: authCheck.status === 403 ? orgKeyRequired : authCheck.error,
+    if (!authCheck.success) {
+      const status = authCheck.error.httpCode;
+      return res.status(status).json({
+        message: status === 403 ? orgKeyRequired : authCheck.error.message,
       });
     }
     // END CHECK AUTH
