+++ web/src/pages/api/public/projects/[projectId]/index.ts
@@ -7,7 +7,7 @@ import {
 } from "@/src/ee/features/admin-api/server/projects/projectById";
 import { hasEntitlementBasedOnPlan } from "@/src/features/entitlements/server/hasEntitlement";
 import { type NextApiRequest, type NextApiResponse } from "next";
-import { verifyOrgAuth } from "@/src/features/auth/policy/verifyOrgAuth";
+import { shadowAuth } from "@/src/features/public-api/server/shadowAuth";
 
 /** orgKeyRequired is the 403 body when a non-organization key hits an organization endpoint. */
 const orgKeyRequired =
@@ -33,13 +33,15 @@ export default async function handler(
   }
 
   // CHECK AUTH
-  const authCheck = await verifyOrgAuth({
+  const authCheck = await shadowAuth({
     req,
     action: req.method === "PUT" ? "project:update" : "project:delete",
+    allowedAccessLevels: ["organization"],
   });
-  if (!authCheck.validKey) {
-    return res.status(authCheck.status).json({
-      message: authCheck.status === 403 ? orgKeyRequired : authCheck.error,
+  if (!authCheck.success) {
+    const status = authCheck.error.httpCode;
+    return res.status(status).json({
+      message: status === 403 ? orgKeyRequired : authCheck.error.message,
     });
   }
   // END CHECK AUTH
