+++ web/src/pages/api/public/projects/[projectId]/memberships/index.ts
@@ -9,7 +9,7 @@ import {
 } from "@/src/ee/features/admin-api/server/projects/projectById/memberships";
 
 import { type NextApiRequest, type NextApiResponse } from "next";
-import { verifyOrgAuth } from "@/src/features/auth/policy/verifyOrgAuth";
+import { shadowAuth } from "@/src/features/public-api/server/shadowAuth";
 
 /** orgKeyRequired is the 403 body when a non-organization key hits an organization endpoint. */
 const orgKeyRequired =
@@ -38,13 +38,15 @@ export default async function handler(
   }
 
   // CHECK AUTH
-  const authCheck = await verifyOrgAuth({
+  const authCheck = await shadowAuth({
     req,
     action: req.method === "GET" ? "projectMembers:read" : "projectMembers:CUD",
+    allowedAccessLevels: ["organization"],
   });
-  if (!authCheck.validKey) {
-    return res.status(authCheck.status).json({
-      error: authCheck.status === 403 ? orgKeyRequired : authCheck.error,
+  if (!authCheck.success) {
+    const status = authCheck.error.httpCode;
+    return res.status(status).json({
+      error: status === 403 ? orgKeyRequired : authCheck.error.message,
     });
   }
   // END CHECK AUTH
