+++ web/src/pages/api/public/scim/ResourceTypes.ts
@@ -1,6 +1,6 @@
 import { cors, runMiddleware } from "@/src/features/public-api/server/cors";
 import { logger } from "@langfuse/shared/src/server";
-import { verifyOrgAuth } from "@/src/features/auth/policy/verifyOrgAuth";
+import { shadowAuth } from "@/src/features/public-api/server/shadowAuth";
 
 import { type NextApiRequest, type NextApiResponse } from "next";
 
@@ -26,15 +26,17 @@ export default async function handler(
   }
 
   // CHECK AUTH
-  const authCheck = await verifyOrgAuth({
+  const authCheck = await shadowAuth({
     req,
     action: "organizationMembers:read",
+    allowedAccessLevels: ["organization"],
   });
-  if (!authCheck.validKey) {
-    return res.status(authCheck.status).json({
+  if (!authCheck.success) {
+    const status = authCheck.error.httpCode;
+    return res.status(status).json({
       schemas: ["urn:ietf:params:scim:api:messages:2.0:Error"],
-      detail: authCheck.status === 403 ? orgKeyRequired : authCheck.error,
-      status: authCheck.status,
+      detail: status === 403 ? orgKeyRequired : authCheck.error.message,
+      status,
     });
   }
   // END CHECK AUTH
